% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{add_asset}
\alias{add_asset}
\title{Create a new Thot Asset with the given properties.
The related data should be saved at the returned path.}
\usage{
add_asset(
  db,
  file,
  name = NULL,
  type = NULL,
  description = NULL,
  tags = NULL,
  metadata = NULL
)
}
\description{
Create a new Thot Asset with the given properties.
The related data should be saved at the returned path.
}
