% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{find_asset}
\alias{find_asset}
\title{Find a single Thot Asset.
If multiple are found, a single one is returned.}
\usage{
find_asset(
  db,
  `_id` = NULL,
  name = NULL,
  type = NULL,
  tags = NULL,
  metadata = NULL
)
}
\description{
Find a single Thot Asset.
If multiple are found, a single one is returned.
}
