% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{find_assets}
\alias{find_assets}
\title{Find Thot Assets matching the filter criteria.}
\usage{
find_assets(
  db,
  `_id` = NULL,
  name = NULL,
  type = NULL,
  tags = NULL,
  metadata = NULL
)
}
\description{
Find Thot Assets matching the filter criteria.
}
