% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{find_containers}
\alias{find_containers}
\title{Find Thot Containers matching the filter criteria.}
\usage{
find_containers(
  db,
  `_id` = NULL,
  name = NULL,
  type = NULL,
  tags = NULL,
  metadata = NULL
)
}
\description{
Find Thot Containers matching the filter criteria.
}
